Rename ResourceElement to ResourceSection
