+++ app/webpacker/components/ResourceSection.js
